refactor(api): extract vitest-free test helpers into @acme/api/testing
The apps/api characterization suite (#660) runs under its own vitest config
and needs these fixtures, but test-utils.ts imports vitest at module scope.
Move everything framework-agnostic into src/testing and re-export it, so the
20 existing test files need no changes.
